Feature(iOS): Rework OAuth proccess in WebView
Hello,

Thank you for providing more information.

Regarding 4.0, your app continues to link out to Safari for GitHub, which provides a poor user experience.

To resolve this issue, it would be appropriate to revise your app to enable log in or registering for an account in the app.

We look forward to reviewing your resubmitted app.

Best regards,

App Store Review